added community button and create screens
