+++ components/Login.jsx
@@ -1,0 +1,14 @@
+import { StyleSheet, Text, View } from 'react-native'
+import React from 'react'
+
+const Login = () => {
+  return (
+    <View>
+      <Text>Login</Text>
+    </View>
+  )
+}
+
+export default Login
+
+const styles = StyleSheet.create({})
